Add click bouncing mechanics with EChests
When you click a enderchest, the click will "bounce" in
all six directions through blocks without hitboxes and
be transmitted to all shop chests and enderchests,
in a recursive manner.

The results will be displayed as if the enderchest was one
shop chest.

To do this, ShopRule was made to be able to be merged
together and contain multiple inventories, by making
each TradeRule refer to an inventory.